Fix refresh of UI
Had the refresh at the bottom of the loop, which meant it waited for
input before refreshing. This was not obvious because I was listening on
startup, which sent an event to the loop that it could process. When I
removed that, suddenly the UI waited for a keypress to draw.