Handle Onion3 listen address
Previously, I had set up the onion stuff so that when it created the
onion service it listened on the local port that Tor will send the
connection to. Now, it listens on an Onion3 address, and that gets
translated in the transport layer to the local listen.

This will allow it to advertise its Onion address properly once I hook
in things like Kademlia.